+++ src/containers/ListingPage/ListingPage.shared.js
@@ -2,10 +2,14 @@ import React from 'react';
 import { FormattedMessage } from '../../util/reactIntl';
 import { types as sdkTypes } from '../../util/sdkLoader';
 import { createResourceLocatorString, findRouteByRouteName } from '../../util/routes';
-import { formatMoney } from '../../util/currency';
+import { convertMoneyToNumber, formatMoney } from '../../util/currency';
 import { timestampToDate } from '../../util/dates';
-import { isUserAuthorized } from '../../util/userHelpers';
-import { NO_ACCESS_PAGE_USER_PENDING_APPROVAL, createSlug } from '../../util/urlHelpers';
+import { hasPermissionToInitiateTransactions, isUserAuthorized } from '../../util/userHelpers';
+import {
+  NO_ACCESS_PAGE_INITIATE_TRANSACTIONS,
+  NO_ACCESS_PAGE_USER_PENDING_APPROVAL,
+  createSlug,
+} from '../../util/urlHelpers';
 
 import { Page, LayoutSingleColumn } from '../../components';
 import FooterContainer from '../../containers/FooterContainer/FooterContainer';
@@ -126,6 +130,10 @@ export const handleContactUser = parameters => () => {
     // A user in pending-approval state can't contact the author (the same applies for a banned user)
     const pathParams = { missingAccessRight: NO_ACCESS_PAGE_USER_PENDING_APPROVAL };
     history.push(createResourceLocatorString('NoAccessPage', routes, pathParams, {}));
+  } else if (!hasPermissionToInitiateTransactions(currentUser)) {
+    // A user in pending-approval state can't contact the author (the same applies for a banned user)
+    const pathParams = { missingAccessRight: NO_ACCESS_PAGE_INITIATE_TRANSACTIONS };
+    history.push(createResourceLocatorString('NoAccessPage', routes, pathParams, {}));
   } else {
     setInquiryModalOpen(true);
   }
@@ -139,6 +147,7 @@ export const handleContactUser = parameters => () => {
  */
 export const handleSubmitInquiry = parameters => values => {
   const { history, params, getListing, onSendInquiry, routes, setInquiryModalOpen } = parameters;
+
   const listingId = new UUID(params.id);
   const listing = getListing(listingId);
   const { message } = values;